Improved conformance for Cards
- Better docs
- Better module names
- Fixed jest problem
- Partly addresses #9
