+++ jest.config.ts
@@ -1,0 +1,9 @@
+import type {Config} from '@jest/types';
+
+const config: Config.InitialOptions = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  rootDir: 'src'
+};
+
+export default config;
